Fixed a null check issue in TryGetAlias
The idea with the method was to specify an alias, and then the language back if the alias existed. However as you may use fenced code blocks without specifying a language (alias), the method would throw an exception. With this commit, the method will instead return false, and "language" will be "None".